backend/requirements.txt
# FastAPI Core
fastapi==0.68.0
uvicorn[standard]==0.15.0

# Database
pymongo==4.6.0
motor==3.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
itsdangerous==2.1.2

# File Processing
PyPDF2==3.0.1
aiofiles==23.2.1

# Utilities
pydantic==1.8.2
python-dotenv==1.0.0

# HTTP & API
httpx==0.25.2
requests==2.31.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
